RAI-752 extract token details route logic (#146)
## Summary
- move token-details response models, SFT detail query structs, aggregate/cache helpers, and route handlers into src/routes/token_details.rs
- keep src/routes/tokens.rs as the route aggregator plus shared token/proof/subgraph helpers
- re-export the details routes so Rocket and utoipa path wiring stay unchanged

## Testing
- nix develop -c cargo fmt
- nix develop -c cargo check
- nix develop -c cargo test routes::tokens::tests
- nix develop -c rainix-rs-static
- local server smoke test: /health, /v1/tokens/details returned 16 rows and 0 errors, /v1/tokens/0x78c31580c97101694c70022c83d570150c11e935/details returned wtSGOV aggregates and activity
- staged local review with two gpt-5.5 reviewers; no Rust behavior findings, only staging hygiene notes addressed

Stacked on #145.
